requirements.txt
torch==2.11.0
torchvision

opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2

albumentations==1.4.8
matplotlib==3.8.4

pillow==10.3.0
tqdm==4.66.1
colorama==0.4.6
pyyaml==6.0
